chore: Extract tags input to separate component
