+++ src/pages/EchoMaker.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import "./EchoMaker.css";
+import EchoMakerTagsInput from "../components/EchoMakerTagsInput";
 
 export default function EchoMaker({ user }) {
   const [formData, setFormData] = useState({
@@ -19,32 +20,10 @@ export default function EchoMaker({ user }) {
   //   }
   // }, []);
 
-  const tags = [
-    "Mood",
-    "Feeling good",
-    "Bad day",
-    "Reflection",
-    "For later",
-    "Note to self",
-    "Motivation",
-    "You and me",
-    "Just us",
-    "Missing you",
-    "Secret",
-  ];
   function createEcho(event) {
     event.preventDefault();
 
     console.log(formData);
-  }
-
-  function removeTag(event) {
-    const tag = event.target.textContent.toLowerCase();
-    const temp = formData.tags.filter((t) => t !== tag);
-    setFormData({
-      ...formData,
-      tags: temp,
-    });
   }
 
   function handleChange(event) {
@@ -93,32 +72,11 @@ export default function EchoMaker({ user }) {
         onChange={handleChange}
         value={formData.description}
       ></textarea>
-      <div>
-        <label htmlFor="tags">Tags:</label>
-        <select name="tags" onChange={handleChange}>
-          {tags.map((tag, index) => (
-            <option
-              key={index}
-              value={tag.toLowerCase()}
-              disabled={formData.tags.includes(tag.toLowerCase())}
-            >
-              {tag}
-            </option>
-          ))}
-        </select>
-        <ul className="list-of-tags">
-          {formData.tags.map((tag, index) => (
-            <li key={tag + index} onClick={removeTag}>
-              {tag
-                .split(" ")
-                .map((word, index) =>
-                  index == 0 ? word[0].toUpperCase() + word.substring(1) : word,
-                )
-                .join(" ")}
-            </li>
-          ))}
-        </ul>
-      </div>
+      <EchoMakerTagsInput
+        formData={formData}
+        setFormData={setFormData}
+        handleChange={handleChange}
+      />
       <label htmlFor="anonymous">
         <input
           type="checkbox"
@@ -126,7 +84,7 @@ export default function EchoMaker({ user }) {
           onChange={handleChange}
           checked={formData.anonymous}
         />
-        Anonymous
+        Post anonymously
       </label>
       <label htmlFor="unlock_datetime">Unlock Time:</label>
       <input
